+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Save on localstore all data. When you request a resource or collection, first check memory. If its empty, read from store. If is empty, get the data from back-end.
 - HttpStorage deprecated: jsons were saved as sent by the server, now we save json with logic (saving ids and resources separately).
 - Service with `toServer()` and `fromServer()` functions. They execute before and after http request. Ideal for type conversions.
+- `JsonapiCore.duplicateResource(resouce, ...relationtypes)` return a duplication of resource. You can duplicate resources and, optionally, their relationships. (v0.6.16)
 
 ## No more declaration file .d.ts